Harmonize use of int (32 or 64) and stop using deprecated numpy aliases
- avoid using deprecated np.int and (-> np.int_) and np.bool (->
  np.bool_)
- use np.int_ instead of np.int32 or np.int64. s dtype. This is
  plateform dependent (correspond to C "long") and is what numpy
  transforms dtype=int into
- use numba.types.int_ instead of numba.types.int32 (same idea as
  np.int_)

+++ discrete_optimization/generic_rcpsp_tools/neighbor_tools_rcpsp.py
@@ -1252,7 +1252,7 @@ class NeighborBuilderMix(NeighborBuilder):
         if isinstance(self.weight_neighbor, list):
             self.weight_neighbor = np.array(self.weight_neighbor)
         self.weight_neighbor = self.weight_neighbor / np.sum(self.weight_neighbor)
-        self.index_np = np.array(range(len(self.list_neighbor)), dtype=np.int32)
+        self.index_np = np.array(range(len(self.list_neighbor)), dtype=np.int_)
         self.verbose = verbose
 
     def find_subtasks(
